requirements.txt
numpy
jupyter
matplotlib
scipy
pandas
h5py
begins
# pymc must be installed separately
